备份一下

+++ rbfx.wasm.apps.server/node-mid/service.js
@@ -536,40 +536,46 @@ function make_graph_for_node(req, res) {
         method: "post",
         url: "http://" + ServerConfig.config.WasmBackstageServer.ip + ":" + ServerConfig.config.WasmBackstageServer.port + "/" + "make_graph_for_node",
         data: json_data,
-    }).then(function (response) {
-        var uuid_str = uuid();
-        var stl_file = "./tmp/" + uuid_str + ".stl";
-        var txt_file = "./tmp/" + uuid_str + ".txt";
-        //
-        if (response.data.length >= 0) {
-            fs.writeFileSync(stl_file, response.data);
+    })
+        .then(function (response) {
+            var uuid_str = uuid();
+            var stl_file = "./tmp/" + uuid_str + ".stl";
+            var txt_file = "./tmp/" + uuid_str + ".txt";
             //
-            var wrcs_ = wasm_rcs();
-            // ----------------------------------------------
-            wrcs_.then((asset) => {
-                var rcs_ = asset;
-                //
-                var ret = rcs_.ccall("DoRun", "int", ["int", "string"], [2, "stl2mdl" + "|" + uuid_str]);
-                if (ret != 0) {
-                    console.log("ResourceToMdlSevice run error.");
-                    res.send("ResourceToMdlSevice error.");
+            if (response.data.length >= 0) {
+                if (response.data != "make_graph_for_node error.") {
+                    fs.writeFileSync(stl_file, response.data);
+                    //
+                    var wrcs_ = wasm_rcs();
+                    // ----------------------------------------------
+                    wrcs_.then((asset) => {
+                        var rcs_ = asset;
+                        //
+                        var ret = rcs_.ccall("DoRun", "int", ["int", "string"], [2, "stl2mdl" + "|" + uuid_str]);
+                        if (ret != 0) {
+                            console.log("ResourceToMdlSevice run error.");
+                            res.send("ResourceToMdlSevice error.");
+                        }
+                        //
+                        var mdl_name = uuid_str + ".mdl";
+                        var mdl_file = "./tmp/" + mdl_name;
+                        const data = fs.readFileSync(mdl_file);
+                        res.send(data);
+                        //
+                        fs.rmSync(mdl_file, { recursive: true, force: true });
+                        fs.rmSync(stl_file, { recursive: true, force: true });
+                        fs.rmSync(txt_file, { recursive: true, force: true });
+                    });
+                } else {
+                    res.send("make_graph_for_node error.");
                 }
-                //
-                var mdl_name = uuid_str + ".mdl";
-                var mdl_file = "./tmp/" + mdl_name;
-                const data = fs.readFileSync(mdl_file);
-                res.send(data);
-                //
-                fs.rmSync(mdl_file, { recursive: true, force: true });
-                fs.rmSync(stl_file, { recursive: true, force: true });
-                fs.rmSync(txt_file, { recursive: true, force: true });
-            });
-        } else {
+            } else {
+                res.send("make_graph_for_node error.");
+            }
+        })
+        .catch(function (error) {
             res.send("make_graph_for_node error.");
-        }
-    }).catch(function (error) {
-        res.send("make_graph_for_node error.");
-    }); 
+        });
 }
 //
 function pbfTogeojsonForSevice(req, res) {
